// src/delivery/enum/filter.enum.ts
export enum FilterCriteria {
  equal = 'equal',
  notEqual = 'notEqual',
  greatherThanOrEqual = 'greatherThanOrEqual',
  lessThanOrEqual = 'lessThanOrEqual',
  greatherThan = 'greatherThan',
  lessThan = 'lessThan',
  includes = 'includes',
  in = 'in',
  contains = 'contains',
  favourites = 'favourites',
  between = 'between',
}